Convert new screen to widgets
