+++ src/screens/Lessons/screens/New/index.js
@@ -1,33 +1,23 @@
 import React from 'react'
 import {
-  newLessonsTitleText,
   newLessonsInstructionsDescriptionText,
   newLessonsInstructionsActionText,
 } from 'utils/text'
 import {newLessonsInfoUrl} from 'utils/urls'
-import Screen from 'components/Screen'
-import Instructions from './components/Instructions'
-import Requested from './components/Requested'
-import Propose from './components/Propose'
+import InstructionsWidget from 'components/InstructionsWidget'
+import RequestedWidget from './components/RequestedWidget'
+import ProposeWidget from './components/ProposeWidget'
 
 export default ({instructor}) => (
-  <Screen
-    title={newLessonsTitleText}
-    intro={
-      instructor.published_lessons === 0 && instructor.pending_lessons === 0
-        ? <Instructions
-            instructor={instructor}
-            description={newLessonsInstructionsDescriptionText}
-            actionText={newLessonsInstructionsActionText}
-            action={newLessonsInfoUrl}
-          />
-        : null
-    }
-    main={
-      <Requested instructor={instructor} />
-    }
-    aside={
-      <Propose instructor={instructor} />
-    }
-  />
+  <div>
+    <InstructionsWidget
+      condition={instructor.published_lessons === 0 && instructor.pending_lessons === 0}
+      instructor={instructor}
+      description={newLessonsInstructionsDescriptionText}
+      actionText={newLessonsInstructionsActionText}
+      action={newLessonsInfoUrl}
+    />
+    <RequestedWidget instructor={instructor} />
+    <ProposeWidget instructor={instructor} />
+  </div>
 )
