+++ src/utils/text.js
@@ -12,6 +12,8 @@ export const newLessonsActionText = 'Create a new lesson'
 export const newLessonsInstructionsDescriptionText = 'You can choose a requested topic below, or propose your own topic. Check out the guide portion on egghead lesson concepts for more details on proposing your own topics.'
 export const newLessonsInstructionsActionText = 'Learn about proposing topics'
 
+export const instructionsTitleText = 'Instructions'
+
 export const lessonsTitleText = 'Lessons'
 
 export const lessonTitleText = 'Lesson'
